+++ src/main/java/com/nvoi/Issuemanagementsystem/service/EventService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface EventService {
     public Event saveEvent(Event event);
-    public List<Event> getEventsByIssueId(int issueId);
+    public List<Event> getEventsByIssueId(long issueId);
 }
